+++ tomni/contour_operations/__init__.py
@@ -1,2 +1,3 @@
 from .get_center import get_center
 from .approximate_circle_by_area import approximate_circle_by_area
+from .circularity import circularity