+++ HISTORY.rst
@@ -3,6 +3,7 @@
 1.12.0 (2022-04-15)
 ------------------
 - BUGFIX: Import approximate_circle_by_area
+- Add contour operation circularity
 
 1.11.0 (2022-04-11)
 ------------------